pyproject.toml
[tool.poetry]
name = "video-clipper"
version = "0.1.0"
description = ""
authors = ["Federico Lanzani <federico@federicolanzani.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
opencv-python = "^4.9.0.80"
streamlit = "^1.35.0"
moviepy = "^1.0.3"
numpy = "^1.26.4"
pillow = "^10.3.0"
streamlit-drawable-canvas = "^0.9.3"
pandas = "^2.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
